create list_id(), rationalise list into spaces
